+++ Sheeeeeeeeet/Appearance/Buttons/ActionSheetButtonAppearance.swift
@@ -1,5 +1,5 @@
 //
-//  ActionSheetButtonItemAppearance.swift
+//  ActionSheetButtonAppearance.swift
 //  Sheeeeeeeeet
 //
 //  Created by Daniel Saidi on 2017-11-24.